--- a/app/javascript/src/component_flow_condition_item.js
+++ b/app/javascript/src/component_flow_condition_item.js
@@ -40,6 +40,11 @@ class FlowConditionItem {
       y2: config.y_out
     };
 
+    this.position = {
+      left: this.bounds.x1,
+      top: this.bounds.y1,
+    }
+
     $node.data("instance", this);
     $node.addClass("FlowConditionItem");
     $node.attr('data-row', this.row);