--- a/app/javascript/src/component_flow_condition_item.js
+++ b/app/javascript/src/component_flow_condition_item.js
@@ -36,6 +36,12 @@ class FlowConditionItem {
     this.$next = config.$next;
     this.row = config.row;
     this.column = config.column;
+    this.bounds = {
+      x1: config.x_in,
+      y1: config.y_in,
+      x2: config.x_out,
+      y2: config.y_out
+    };
   }
 }