--- a/app/javascript/src/component_flow_item.js
+++ b/app/javascript/src/component_flow_item.js
@@ -15,7 +15,6 @@
 
 
 
-
 /* FlowItem component
  * ------------------------
  * Positionable item expected within a Flow Layout.
@@ -33,10 +32,11 @@ class FlowItem {
     this.next = config.next;
     this.row = config.row;
     this.column = config.column;
-    this.coords = {
-      x_in: config.x_in,
-      x_out: config.x_out,
-      y: config.y,
+    this.bounds = {
+      x1: config.x_in,
+      y1: config.y_in,
+      x2: config.x_out,
+      y2: config.y_out
     };
 
   }